NPC: Moved NpcEventBus to separate file
